fix native build / type checking
